Moved hasher settings to separate flie
